Separated Model definition and table creation into two files
